Move helper function to utils
